Revert "A little bit of refactoring and division into blocks"
