+++ .gitignore
@@ -5,4 +5,3 @@
 *.db
 test.py
 /downloads/
-.idea
