--- config/app_config.py
@@ -1,4 +1,0 @@
-ver_app = "v 0.1.1.1 beta"
-name_app = "Forward bot by: @SYSdeppord "
-device_model = "SYSdeppord govno cloud"
-system_version = "Linux chroot on Android"
