--- config/tg_config.py
@@ -1,6 +1,0 @@
-import os
-
-api_id = os.environ["api_id"]
-api_hash = os.environ["api_hash"]
-bot_token = os.environ["bot_token"]
-admin_id = int(os.environ["admin_id"])
