+++ main.py
@@ -1,23 +1,22 @@
-# import os
+import os
 from pyrogram import Client, filters, idle
 from pyrogram.handlers import MessageHandler
-from cp_bot.bot_processor import Sorter
-from handlers.handlers import UserHandlers
-from config.app_config import *
-from config.tg_config import *
-from proxy_class import setting
+from tg_config import api_id, api_hash, bot_token, setting, name_app, ver_app, system_version, device_model
+from bot_processor import Sorter
+from handlers import UserHandlers
 
 bot = Client("bot", api_id=api_id, api_hash=api_hash, bot_token=bot_token, workdir="./files/bot")
-
 users = []
 
 
-def add_user_client():
+def build_user_apps():
     print("Building users apps")
     for usr in setting.user_setting:
         if setting.user_setting[f"{usr}"]["authorised"] and os.path.exists(f"./files/users/u{usr}"):
             user_id = usr
             name = f"u{user_id}"
+            #if not os.path.exists(f"./files/users/{name}"):
+            #    os.mkdir(f"./files/users/{name}")
             users.append(Client(name, api_id=api_id, api_hash=api_hash, app_version=name_app + ver_app,
                                 device_model=device_model, system_version=system_version,
                                 workdir=f"./files/users/{name}"))
@@ -48,7 +47,7 @@ user_handlers = UserHandlers(bot)
 user_message = user_handlers.user_message
 
 if __name__ == "__main__":
-    add_user_client()
+    build_user_apps()
     check_and_create_folders()
     print("Starting control panel")
     bot.start()
