--- proxy_class.py
@@ -1,6 +1,0 @@
-from database.db_engine import Setting
-# from user_bot.user_clients import UserClients
-
-setting = Setting()
-#user_clients = UserClients()
-#users = user_clients.users
