+++ tg_config.py
@@ -1,0 +1,12 @@
+import os
+from db_engine import Setting
+
+setting = Setting()
+api_id = os.environ["api_id"]
+api_hash = os.environ["api_hash"]
+bot_token = os.environ["bot_token"]
+admin_id = int(os.environ["admin_id"])
+ver_app = "v 0.1.1.1 beta"
+name_app = "Forward bot by: @SYSdeppord "
+device_model = "SYSdeppord govno cloud"
+system_version = "Linux chroot on Android"
